--- three/src/main/java/com/wizardlybump17/physics/three/group/AbstractObjectsGroup.java
@@ -1,75 +1,0 @@
-package com.wizardlybump17.physics.three.group;
-
-import com.wizardlybump17.physics.three.container.BaseObjectContainer;
-import com.wizardlybump17.physics.three.object.BaseObject;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicInteger;
-
-public abstract class AbstractObjectsGroup implements ObjectsGroup {
-
-    private static final @NotNull AtomicInteger BODY_COUNTER = new AtomicInteger(0); //what is your body count?
-
-    private final int id = BODY_COUNTER.getAndIncrement();
-    private final @NotNull BaseObjectContainer container;
-    private final @NotNull Map<Integer, BaseObject> objects;
-
-    public AbstractObjectsGroup(@NotNull BaseObjectContainer container, @NotNull Collection<BaseObject> objects) {
-        this.container = container;
-
-        Map<Integer, BaseObject> internalObjects = new HashMap<>();
-        for (BaseObject object : objects)
-            internalObjects.put(object.getId(), object);
-        this.objects = Collections.unmodifiableMap(internalObjects);
-    }
-
-    @Override
-    public int getId() {
-        return id;
-    }
-
-    @Override
-    public @NotNull BaseObjectContainer getContainer() {
-        return container;
-    }
-
-    @Override
-    public @NotNull Map<Integer, BaseObject> getObjects() {
-        return objects;
-    }
-
-    @Override
-    public boolean hasObject(int objectId) {
-        return objects.containsKey(objectId);
-    }
-
-    @Override
-    public abstract boolean isPassable();
-
-    @Override
-    public abstract boolean isCollidingWith(@NotNull BaseObject other);
-
-    @Override
-    public void tick() {
-        tickCollisions();
-    }
-
-    protected void tickCollisions() {
-        for (ObjectsGroup otherGroup : getContainer().getObjectsGroups()) {
-            if (id == otherGroup.getId())
-                continue;
-
-            if (isCollidingWith(otherGroup)) {
-                onCollide(otherGroup);
-                otherGroup.onCollide(this);
-            } else {
-                onStopColliding(otherGroup);
-                otherGroup.onStopColliding(this);
-            }
-        }
-    }
-}
