+++ three/src/main/java/com/wizardlybump17/physics/three/group/PhysicsObjectsGroup.java
@@ -1,6 +1,5 @@
 package com.wizardlybump17.physics.three.group;
 
-import com.wizardlybump17.physics.Constants;
 import com.wizardlybump17.physics.three.Vector3D;
 import com.wizardlybump17.physics.three.container.BaseObjectContainer;
 import com.wizardlybump17.physics.three.object.BaseObject;
@@ -11,7 +10,7 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.Set;
 
-public abstract class PhysicsObjectsGroup extends AbstractObjectsGroup {
+public abstract class PhysicsObjectsGroup extends ObjectsGroup {
 
     private final @NotNull Set<Integer> collidingWith = new HashSet<>();
     private @NotNull Vector3D acceleration;
@@ -41,7 +40,7 @@ public abstract class PhysicsObjectsGroup extends AbstractObjectsGroup {
     }
 
     @Override
-    public void onCollide(@NotNull ObjectsGroup otherGroup) {
+    protected void onCollide(@NotNull ObjectsGroup otherGroup) {
         collidingWith.add(otherGroup.getId());
 
         setAcceleration(Vector3D.ZERO);
@@ -49,22 +48,34 @@ public abstract class PhysicsObjectsGroup extends AbstractObjectsGroup {
     }
 
     @Override
-    public void onStopColliding(@NotNull ObjectsGroup otherGroup) {
+    protected void onStopColliding(@NotNull ObjectsGroup otherGroup) {
         collidingWith.remove(otherGroup.getId());
     }
 
+    /**
+     * @return the acceleration of this group, in meters per tick
+     */
     public @NotNull Vector3D getAcceleration() {
         return acceleration;
     }
 
+    /**
+     * @param acceleration the acceleration to set, in meters per tick
+     */
     public void setAcceleration(@NotNull Vector3D acceleration) {
         this.acceleration = acceleration;
     }
 
+    /**
+    * @return the velocity of this group, in meters per tick
+    */
     public @NotNull Vector3D getVelocity() {
         return velocity;
     }
 
+    /**
+    * @param velocity the velocity to set, in meters per tick
+    */
     public void setVelocity(@NotNull Vector3D velocity) {
         this.velocity = velocity;
     }
@@ -76,7 +87,24 @@ public abstract class PhysicsObjectsGroup extends AbstractObjectsGroup {
     }
 
     protected void tickMovement() {
-        setVelocity(getVelocity().add(getAcceleration().divide(Constants.TICKS_PER_SECOND)));
-        setCenter(getCenter().add(getVelocity().divide(Constants.TICKS_PER_SECOND)));
+        velocity = getMaxMovement(velocity.add(acceleration));
+        setCenter(getCenter().add(velocity));
+    }
+
+    /**
+     * <p>
+     * Returns the maximum movement this object group can reach given the original desired movement.
+     * An example of this, is when the original movement would cause the object group to go past the ground,
+     * so the returned value would be the movement that would keep the object group on the ground.
+     * </p>
+     * <p>
+     * The returned value must be in meters per tick.
+     * </p>
+     *
+     * @param movement the original desired movement
+     * @return the maximum movement this object group can reach given the original desired movement.
+     */
+    public @NotNull Vector3D getMaxMovement(@NotNull Vector3D movement) {
+        return movement;
     }
 }
